requirements.txt
requests==2.32.3
kafka-python==2.0.6
pyspark==3.5.0
transformers==4.37.2
torch==2.2.0
cassandra-driver==3.28.0
nltk==3.8.1
pandas==2.2.0 